fix: complete int32_t type migration for transaction sizes
Add missing int32_t change in UpdateEntryForAncestors updateSize variable.
This aligns with Bitcoin PR #23962 intent to standardize transaction size types
and addresses the validation finding that changes were incomplete.

- Change int64_t updateSize to int32_t updateSize in UpdateEntryForAncestors
- Brings change ratio to proper alignment with Bitcoin commit scope
- Completes type consistency improvements from original Bitcoin commit

+++ src/txmempool.cpp
@@ -332,7 +332,7 @@ void CTxMemPool::UpdateAncestorsOf(bool add, txiter it, setEntries &setAncestors
 void CTxMemPool::UpdateEntryForAncestors(txiter it, const setEntries &setAncestors)
 {
     int64_t updateCount = setAncestors.size();
-    int64_t updateSize = 0;
+    int32_t updateSize = 0;
     CAmount updateFee = 0;
     int updateSigOps = 0;
     for (txiter ancestorIt : setAncestors) {
